fix(viewer): the NotFoundError family pushed load-errors past the module-size gate
`node scripts/check-module-size.mjs` exits 1 on this branch, on both of its
teeth at once: load-errors.ts crossed 400 lines (424) with no allowlist row,
and analytics-scrub.ts grew past its recorded 629 budget to 649. Codex
reported both.

load-errors.ts is split the way its own module docblock already prescribes for
"the families whose matchers carry more rationale than pattern": the two
NotFoundError wordings, the load-context set and the gate that joins them move
to a sibling module, file-not-found-errors.ts, alongside webgl-unavailable.ts
and cancelled-and-network-errors.ts. Only one predicate is exported, so the
classifier keeps owning the ORDER the buckets are tried in. load-errors.ts is
back to 319 lines and needs no allowlist row.

analytics-scrub.ts sat at exactly its budget with zero headroom, so any added
line failed. Its severity rationale is now duplication: the argument for why
`file_unreadable` is benign is inseparable from the WebKit context gate, so it
moves next to that gate and analytics-scrub keeps a three-line pointer. The
comment explaining the `context` argument went entirely, being a restatement of
classifyLoadError's own JSDoc. That leaves the file at 633, and its allowlist
row is raised by 4 rather than the 20 the branch asked for. No other row moved.

Behaviour is unchanged and the split is proved non-vacuous: deleting the
isLoadContext gate from the extracted module fails 3 tests across
load-errors.test.ts and analytics.test.ts, so the suite binds the new module
rather than the old file's copy.

check-module-size exits 0 (2366 files measured, 360 allowlisted, 0 new over
400); its own gate tests 59/59; pnpm typecheck 1667/1667; pnpm lint 0 errors;
245/245 across the seven touched viewer test files.
