+++ apps/viewer/src/lib/load-error-message.ts
@@ -26,8 +26,8 @@ import {
  *
  * @param fileName Optional file name to attribute the failure to.
  * @param context  The capture site's `context`, where the caller has one. Only
- *   one WebKit wording depends on it (see `isWebkitGenericNotFoundMessage` in
- *   ./load-errors.ts); omitting it costs that string its guidance and nothing
+ *   one WebKit wording depends on it (see `isFileNotFoundMessage` in
+ *   ./file-not-found-errors.ts); omitting it costs that string its guidance and nothing
  *   else. A caller that is NOT a model load should omit it rather than invent
  *   one, because that is precisely the case the gate exists to exclude.
  */
