+++ scripts/module-size-allowlist.txt
@@ -131,7 +131,7 @@
    436 apps/viewer/src/hooks/useIfcServer.ts
    488 apps/viewer/src/hooks/useKeyboardShortcuts.ts
    496 apps/viewer/src/hooks/useSymbolicAnnotations.ts
-   629 apps/viewer/src/lib/analytics-scrub.ts
+   633 apps/viewer/src/lib/analytics-scrub.ts
    651 apps/viewer/src/lib/clash/persistence.ts
    427 apps/viewer/src/lib/collab/geometry-sync.ts
    445 apps/viewer/src/lib/compare/buildFingerprints.ts
